Remove test for viewing add-on source code.
As detailed in https://github.com/mozilla/addons-server/issues/5135 the view source feature was disabled due to performance issues. It looks unlikely to be reintroduced, and as we are intending to retire this repository once sufficient tests are moved to the addons-server repository, it makes sense to remove this test.